requirements.txt
# ASCENSÃO DO CULTIVO DIGITAL - Dependências Atualizadas (Dez 2025)
# Versão 3.0 com ferramentas de segurança e hacking atualizadas

# Backend - FastAPI
fastapi==0.123.9
uvicorn[standard]==0.34.0
python-multipart==0.0.7
python-socketio==5.11.0
python-engineio==4.9.0

# Captura e Análise de Tráfego
scapy==2.6.1
psutil==6.1.1
netifaces2==0.0.22
mitmproxy==11.0.0

# Processamento de Dados
pandas==2.2.3
numpy==1.26.4 # (Mantido na v1 para evitar breaking changes da v2)

# Geração de Relatórios
reportlab==4.2.5
fpdf2==2.8.2

# Utilitários Web e HTTP
requests==2.32.3
aiofiles==24.1.0
pydantic==2.10.5
pydantic-settings==2.7.0

# Ferramentas de Segurança e Hacking
beautifulsoup4==4.14.3
selenium==4.27.0
playwright==1.49.0
requests-html==0.10.0
python-nmap==0.7.1
paramiko==3.5.0
PySocks==1.7.1
yara-python==4.5.1
webdriver-manager==4.0.2
lxml[html_clean]==5.3.0

# Desenvolvimento (opcional)
pytest==8.3.4
pytest-asyncio==0.23.0
black==24.10.0
flake8==7.1.1
